refactored most of e2e tests
