+++ supportWebApp/src/app/components/ticket-view/ticket-view.component.ts
@@ -8,67 +8,75 @@ import { map, catchError } from 'rxjs/operators';
 import { Observable } from 'rxjs/Observable';
 
 @Component({
-  selector: 'app-ticket-view',
-  templateUrl: './ticket-view.component.html',
-  styleUrls: ['./ticket-view.component.css']
+    selector: 'app-ticket-view',
+    templateUrl: './ticket-view.component.html',
+    styleUrls: ['./ticket-view.component.css']
 })
 export class TicketViewComponent implements OnInit, AfterViewInit {
-  dataSource: MatTableDataSource<Ticket> = new MatTableDataSource();
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
-  displayedColumns = [
-    'status',
-    'priority',
-    'subject',
-    'requester',
-    'assignee',
-    'updatedAt',
-    'gototick'
-  ];
-  public criticalTickets: number;
-  public openTickets: number;
+    dataSource: MatTableDataSource<Ticket> = new MatTableDataSource();
+    @ViewChild(MatPaginator) paginator: MatPaginator;
+    @ViewChild(MatSort) sort: MatSort;
+    displayedColumns = [
+        'status',
+        'priority',
+        'subject',
+        'requester',
+        'assignee',
+        'updatedAt',
+        'gototick'
+    ];
+    public criticalTickets: number;
+    public openTickets: number;
 
-  constructor(
-    private ticketService: TicketService,
-    private dataservice: DataService
-  ) {
-    this.getCriticalAmount();
-    this.getOpenAmount();
-  }
+    constructor(
+        private ticketService: TicketService,
+        private dataservice: DataService
+    ) {
+        this.getCriticalAmount();
+        this.getOpenAmount();
+    }
 
-  getOpenAmount() {
-    this.ticketService.getOpenAmount().subscribe(result => {
-      this.openTickets = result as number;
-    });
-  }
+    getOpenAmount() {
+        this.ticketService.getOpenAmount().subscribe(result => {
+            if (result != null) {
+                this.openTickets = result as number;
+            } else {
+                this.openTickets = 0;
+            }
+        });
+    }
 
-  getCriticalAmount() {
-    this.ticketService.getcriticalAmount().subscribe(result => {
-      this.criticalTickets = result as number;
-    });
-  }
+    getCriticalAmount() {
+        this.ticketService.getcriticalAmount().subscribe(result => {
+            if (result != null) {
+                this.criticalTickets = result as number;
+            } else {
+                this.criticalTickets = 0;
+            }
+        });
+    }
 
-  getTickets() {
-    this.ticketService.getTickets().subscribe(result => {
-      this.dataSource.data = result as Ticket[];
-    });
-  }
+    getTickets() {
+        this.ticketService.getTickets().subscribe(result => {
+            this.dataSource.data = result as Ticket[];
+        });
+    }
 
-  goToTicket(ticket: Ticket) {
-    this.dataservice.changeCurrentTicket(ticket);
-  }
+    goToTicket(ticket: Ticket) {
+        this.dataservice.changeCurrentTicket(ticket);
+    }
 
-  ngAfterViewInit() {
-    this.dataSource.paginator = this.paginator;
-    this.dataSource.sort = this.sort;
-    this.sort.sortChange.subscribe(() => (this.paginator.pageIndex = 0));
-    this.getTickets();
-  }
+    ngAfterViewInit() {
+        this.dataSource.paginator = this.paginator;
+        this.dataSource.sort = this.sort;
+        this.sort.sortChange.subscribe(() => (this.paginator.pageIndex = 0));
+        this.getTickets();
+    }
 
-  applyFilter(filterValue: string) {
-    filterValue = filterValue.trim();
-    filterValue = filterValue.toLowerCase();
-    this.dataSource.filter = filterValue;
-  }
-  ngOnInit() {}
+    applyFilter(filterValue: string) {
+        filterValue = filterValue.trim();
+        filterValue = filterValue.toLowerCase();
+        this.dataSource.filter = filterValue;
+    }
+    ngOnInit() { }
 }
